tweaking with the UI/UX

+++ app/src/main/java/com/blarknes/doransvault/DAO/DataBaseHelper.java
@@ -2,16 +2,17 @@ package com.blarknes.doransvault.DAO;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
-
 import com.blarknes.doransvault.model.Conta;
 import com.j256.ormlite.android.apptools.OrmLiteSqliteOpenHelper;
 import com.j256.ormlite.support.ConnectionSource;
 import com.j256.ormlite.table.TableUtils;
 
 public class DataBaseHelper extends OrmLiteSqliteOpenHelper {
+    private static final String databaseName = "AccountsDatabase.db";
+    private static final int databaseVersion = 4;
 
     public DataBaseHelper(Context context) {
-        super(context, "sistema.db", null, 3);
+        super(context, databaseName, null, databaseVersion);
     }
 
     @Override
@@ -26,8 +27,10 @@ public class DataBaseHelper extends OrmLiteSqliteOpenHelper {
     @Override
     public void onUpgrade(SQLiteDatabase db, ConnectionSource src, int oldVersion, int newVersion) {
         try{
-            TableUtils.dropTable(src, Conta.class, true);
-            onCreate(db, src);
+            if (oldVersion != newVersion) {
+                TableUtils.dropTable(src, Conta.class, true);
+                onCreate(db, src);
+            }
         }catch(Exception e){
             e.printStackTrace();
         }
